feat: 슬라이더 수정(#226)
